+++ cmd/rancid-run/main.go
@@ -12,7 +12,7 @@ import (
 	"gorancid/pkg/config"
 )
 
-const version = "0.4.1"
+const version = "0.4.2"
 
 func main() {
 	var (